Cleanup specfile for rpmlint
This adresses the errors rpmlint was seeing. Most of them where not critical since this package has rather limited use as of now. There may still be some warnings but I'm not fixing those since I do not see any value in changing something that works (and was acutally hard to figure out, damn configure and make args).

Fixes #5